🔒 Warden: [security fix] Fix timing leak in HMAC verification
